+++ src/Products/PloneMeeting/browser/overrides.py
@@ -1327,8 +1327,7 @@ class PMDocumentGenerationView(DashboardDocumentGenerationView):
         """Generates the stored annex title using the ConfigurablePODTemplate.store_as_annex_title_expr.
            If empty, we just return the ConfigurablePODTemplate title."""
         value = pod_template.store_as_annex_title_expr
-        extra_expr_ctx = _base_extra_expr_ctx(
-            self.context, {'obj': self.context, 'pod_template': pod_template})
+        extra_expr_ctx = _base_extra_expr_ctx(self.context, {'pod_template': pod_template})
         evaluatedExpr = _evaluateExpression(
             self.context,
             expression=value and value.strip() or '',